trunk: Added more checks to the Android package builder so that it actual quits when the build fails.
git-svn-id: svn://localhost/public/MoSync/trunk@1967 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ runtimes/java/platforms/androidJNI/buildJava.rb
@@ -35,11 +35,19 @@ def preprocess_android_file(src_file, src_dir, shared_dir, output_dir)
 	end 
 	
 	# Preprocess the jpp file into a jtmp file, sed fixes the output if any
-	system("xgcc -x c -E -o #{output_dir}#{jtmp_file} #{buildFlags} -I#{shared_dir} -Isrc" +
+	success = system("xgcc -x c -E -o #{output_dir}#{jtmp_file} #{buildFlags} -I#{shared_dir} -Isrc" +
 		" #{src_dir}#{src_file} 2>&1 | sed \"s/\\([a-zA-Z/]\\+\\)\\(.[a-zA-Z]\\+\\):\\([0-9]\\+\\):/\\1\\2(\\3):/\"")
 	
+	if (!success)
+		exit 1
+	end
+	
 	# Use sed to comment the lines which the proprocessor added to the file and save it as a java file
-	system("sed \"s/^# /\\/\\//\" < #{output_dir}#{jtmp_file} > #{output_dir}#{java_file}");
+	success = system("sed \"s/^# /\\/\\//\" < #{output_dir}#{jtmp_file} > #{output_dir}#{java_file}");
+	
+	if (!success)
+		exit 1
+	end
 	
 	FileUtils.remove "#{output_dir}#{jtmp_file}"
 end
@@ -76,3 +84,5 @@ Dir.foreach(android_source) {|x|
 		preprocess_android_file(x, "#{android_source}/", shared_java_source, out_dir)
 	end
 }
+
+exit 0
